Add support to save images to cloudinary for tournaments + general solution
The image class now supports full uploads to cloundinary. This will save the image and return the cloudinary response, which has a url that can be saved in the database for later acess. Updated the tournament create logic to support the cloudinary uploads so in the future for any usecase this can be an example of how to use the image + cloudinary class.
To access cloudinary's images `heroku addons:open cloudinary` on CLI