Add Support for oraganizer dashboard to see all user information
All users who are part of the tournament can see information regarding the users status, as well as send them email, or check them in. The buttons for checking in the user and sending them an email still need to be made functional. At the momment they are placeholders for the view