Org Dash views updated to reflect admin privalage
Admins will be able to view all users and see their status of if their admins or not. They are NOT allowed to set the admin status of any other user. Only the organizer will have this.

Trello: [od] Admins of a tournament see subset of data on organizer dashboard