Add support for super admin access
If a user is an admin (users table is_admin = true), then in their my accounts section they will be able to auto login as any other user. This is going to be beneficial incase certain bugs are only prevalent to specific users, now we will be able to check what the bug actaully is and how their seeing it.

Overall this change allows admins to auto login as any user